+++ NEWS.md
@@ -3,6 +3,7 @@
 - Fixed: `newProject()` would not add suspect annotation to the output script if the example suspect list or sets workflows were chosen.
 - Fixed: Default optimization range for KPIC2 `min_width` was incorrect (PR #31, thanks to @@coltonlloyd)
 - `installPatRoon()` improvements in determining what is already installed
+- Fixed: Group qualities/scores were not transferred to new `featureGroups` objects after calling `screenSuspects()` or `unset()`
 
 
 # patRoon 2.0